+++ src/test/java/com/maxmind/geoip2/WebServiceClientTest.java
@@ -116,7 +116,7 @@ public class WebServiceClientTest {
         assertNull(traits.getIsp());
         assertNull(traits.getOrganization());
         assertNull(traits.getUserType());
-        assertNull(traits.getStaticIPScore());
+        assertNull(traits.getStaticIpScore());
         assertNull(traits.getUserCount());
         assertFalse(traits.isAnonymousProxy());
         assertFalse(traits.isSatelliteProvider());